kerchunk.py -> types.kerchunk.py
