+++ virtualizarr/types/general.py
