+++ virtualizarr/types/kerchunk.py
@@ -1,0 +1,12 @@
+from typing import NewType
+
+# Distinguishing these via type hints makes it a lot easier to mentally keep track of what the opaque kerchunk "reference dicts" actually mean
+# (idea from https://kobzol.github.io/rust/python/2023/05/20/writing-python-like-its-rust.html)
+# TODO I would prefer to be more specific about these types
+KerchunkStoreRefs = NewType(
+    "KerchunkStoreRefs", dict
+)  # top-level dict with keys for 'version', 'refs'
+KerchunkArrRefs = NewType(
+    "KerchunkArrRefs",
+    dict,
+)  # lower-level dict containing just the information for one zarr array
